rename main GUI classes
