+++ src/main/java/com/ateam/onpoint/gui/WindowPane.java
@@ -1,0 +1,21 @@
+package com.ateam.onpoint.gui;
+
+
+import javafx.scene.layout.AnchorPane;
+import javafx.scene.layout.StackPane;
+
+/**
+ * Manages all the components specific to the application
+ */
+public class WindowPane extends AnchorPane {
+    public static final int MIN_WINDOW_WIDTH = 840;
+    public static final int MIN_WINDOW_HEIGHT = 420;
+    public static final int SIDEBAR_WIDTH = 200;
+
+    public WindowPane() {
+        StackPane body = new StackPane(new Dashboard(new OnPointGUI()));
+        body.getStyleClass().add("body");
+
+        this.getChildren().add(body);
+    }
+}
