+++ src/main/java/com/ateam/onpoint/gui/components/NavTree.java
@@ -1,6 +1,6 @@
 package com.ateam.onpoint.gui.components;
 
-import com.ateam.onpoint.gui.ApplicationView;
+import com.ateam.onpoint.gui.OnPointGUI;
 import com.ateam.onpoint.gui.content.IContent;
 import javafx.geometry.Pos;
 import javafx.scene.control.Label;
@@ -17,7 +17,7 @@ public class NavTree extends TreeView<NavTree.ContentRecord> {
      * Builds the NavTree and the factor for the cell constructor.
      * @param view
      */
-    public NavTree(ApplicationView view) {
+    public NavTree(OnPointGUI view) {
         super();
 
         this.setShowRoot(false);
